ep3.py
This is a pure python program written for line clipping using both the algorithms: Cohen-Sutherland and Liang-Barsky. The "beyond approach" is that you can have the analysis of both the results run in different terminals(command line interface) at the same time, while commenting each algorithm at once, in one window of the two used.